moving images to images folder
