+++ src/components/homePage-components/hero-section/hero.module.css
@@ -7,7 +7,7 @@
   align-items: center;
   justify-content: center;
   height: 80vh;
-  background-image: url("../../../../public/hero-bg.png"); /* Replace with your image path */
+  background-image: url("../../../images/hero-bg.png"); /* Replace with your image path */
   background-size: cover;
   background-position: center;
   color: white;
@@ -26,7 +26,7 @@
     0,
     0,
     0,
-    0.8
+    0.7
   ); /* Dark overlay for better text readability */
 }
